+++ app/views/applications/process/incomes/index.html.slim
@@ -1,4 +1,4 @@
-= form_for @form, as: :application, url: application_income_save_path, method: :put, html: { autocomplete: 'off' } do |f|
+= form_for @form, as: :application, url: application_incomes_path, method: :post, html: { autocomplete: 'off' } do |f|
 
   .grid-row
     .column-two-thirds
