+++ spec/controllers/applications/process_controller_spec.rb
@@ -3,16 +3,13 @@ require 'rails_helper'
 RSpec.describe Applications::ProcessController, type: :controller do
   let(:user)          { create :user }
   let(:application) { build_stubbed(:application, office: user.office) }
-  let(:income_form) { instance_double('Forms::Application::Income') }
-  let(:income_calculation_runner) { instance_double(IncomeCalculationRunner, run: nil) }
   let(:dwp_monitor) { instance_double('DwpMonitor') }
   let(:dwp_state) { 'online' }
 
   before do
     sign_in user
     allow(Application).to receive(:find).with(application.id.to_s).and_return(application)
-    allow(Forms::Application::Income).to receive(:new).with(application).and_return(income_form)
-    allow(IncomeCalculationRunner).to receive(:new).with(application).and_return(income_calculation_runner)
+
     allow(dwp_monitor).to receive(:state).and_return(dwp_state)
     allow(DwpMonitor).to receive(:new).and_return(dwp_monitor)
   end
@@ -33,73 +30,6 @@ RSpec.describe Applications::ProcessController, type: :controller do
 
     it 'redirects to the personal information page for that application' do
       expect(response).to redirect_to(application_personal_informations_path(application))
-    end
-  end
-
-  describe 'GET #income' do
-    let(:application) { build_stubbed(:application, office: user.office, benefits: benefits) }
-
-    before do
-      get :income, application_id: application.id
-    end
-
-    context 'when application is on benefits' do
-      let(:benefits) { true }
-
-      it 'redirects to the summary' do
-        expect(response).to redirect_to(application_summary_path(application))
-      end
-    end
-
-    context 'when application is not on benefits' do
-      let(:benefits) { false }
-
-      it 'returns 200 response' do
-        expect(response).to have_http_status(200)
-      end
-
-      it 'renders the correct template' do
-        expect(response).to render_template(:income)
-      end
-
-      it 'assigns the income form' do
-        expect(assigns(:form)).to eql(income_form)
-      end
-    end
-  end
-
-  describe 'PUT #income_save' do
-    let(:expected_params) { { dependents: false } }
-
-    before do
-      allow(income_form).to receive(:update_attributes).with(expected_params)
-      allow(income_form).to receive(:save).and_return(form_save)
-
-      put :income_save, application_id: application.id, application: expected_params
-    end
-
-    context 'when the form can be saved' do
-      let(:form_save) { true }
-
-      it 'runs the income calculation on the application' do
-        expect(income_calculation_runner).to have_received(:run)
-      end
-
-      it 'redirects to the summary page' do
-        expect(response).to redirect_to(application_summary_path(application))
-      end
-    end
-
-    context 'when the form can\'t be saved' do
-      let(:form_save) { false }
-
-      it 'renders the correct template' do
-        expect(response).to render_template(:income)
-      end
-
-      it 'assigns the income form' do
-        expect(assigns(:form)).to eql(income_form)
-      end
     end
   end
